+++ supplier_api_tester/supplier_api_tester/tests/reservation.py
@@ -21,7 +21,7 @@ def test_missing_api_key(api_url, api_key, product_id, timeslots: bool, version=
 
     if raw_response.status_code != 403:
         raise FailedTest(
-            message=f'Incorrect status code ({raw_response.status_code}) when calling the API wihout the API-Key. Expected status code: 403.',
+            message=f'Incorrect status code "{raw_response.status_code}" when calling the API wihout the API-Key. Expected status code: "403".',
             response=raw_response,
         )
 
@@ -29,8 +29,8 @@ def test_missing_api_key(api_url, api_key, product_id, timeslots: bool, version=
         return TestResult(
             status=1,
             message=(
-                f'Incorrect text message ({raw_response.text}). '
-                'Expected message: Forbidden - Missing or incorrect API key'
+                f'Incorrect text message "{raw_response.text}". '
+                'Expected message: "Forbidden - Missing or incorrect API key".'
             )
         )
 
@@ -47,7 +47,7 @@ def test_incorrect_api_key(api_url, api_key, product_id, timeslots: bool, versio
 
     if raw_response.status_code != 403:
         raise FailedTest(
-            message=f'Incorrect status code ({raw_response.status_code}) when calling the API wihout the API-Key. Expected status code: 403.',
+            message=f'Incorrect status code "{raw_response.status_code}" when calling the API wihout the API-Key. Expected status code: "403".',
             response=raw_response,
         )
 
@@ -55,8 +55,8 @@ def test_incorrect_api_key(api_url, api_key, product_id, timeslots: bool, versio
         return TestResult(
             status=1,
             message=(
-                f'Incorrect text message ({raw_response.text}). '
-                'Expected message: Forbidden - Missing or incorrect API key'
+                f'Incorrect text message "{raw_response.text}". '
+                'Expected message: "Forbidden - Missing or incorrect API key".'
             )
         )
 
@@ -77,7 +77,7 @@ def test_missing_argument_error(api_url, api_key, product_id, timeslots: bool, v
     expected_error = ApiError(
         error_code=1000,
         error='Missing argument',
-        message="Required argument date was not found",
+        message='Required argument "date" was not found',
     )
     result = check_api_error(raw_response, api_error, expected_error)
     if result.is_warning:
@@ -90,7 +90,7 @@ def test_missing_argument_error(api_url, api_key, product_id, timeslots: bool, v
     expected_error = ApiError(
         error_code=1000,
         error='Missing argument',
-        message="Required argument tickets was not found",
+        message='Required argument "tickets" was not found',
     )
     result = check_api_error(raw_response, api_error, expected_error)
     if result.is_warning:
@@ -106,7 +106,7 @@ def test_missing_argument_error(api_url, api_key, product_id, timeslots: bool, v
     expected_error = ApiError(
         error_code=1000,
         error='Missing argument',
-        message="Required argument customer was not found",
+        message='Required argument "customer" was not found',
     )
     result = check_api_error(raw_response, api_error, expected_error)
     if result.is_warning:
@@ -194,7 +194,7 @@ def test_not_allowed_method(api_url, api_key, product_id, timeslots: bool, versi
         status_code = getattr(raw_response, 'status_code', 200)
         if status_code != 405:
             raise FailedTest(
-                message=f'Incorrect status code ({status_code}) when calling the API via method {method.__name__.upper()}. Expected status code: 405.',
+                message=f'Incorrect status code "{status_code}" when calling the API via method {method.__name__.upper()}. Expected status code: "405".',
                 response=raw_response,
             )
     return TestResult()